cleanup, rename buttons > bordered/borderless
